FREEPBX-23203: Fix error "Trying to access array offset on value of type bool".
Remove user "Template Creator".
Define the user's displayname "Template Creator" with a private variable of the object.
Define the user "FreePBXUCPTemplateCreator" with a private variable of the object.
Change the validate mode if the user is "FreePBXUCPTemplateCreator". Now instead of using displayname we use the real name of the user.